backend/requirements.txt
Django==5.0.*
djangorestframework==3.14.*
psycopg2-binary==2.9.*
redis==5.0.0
django-redis==5.*
celery[redis]==5.3.*
django-environ==0.11.*
django-cors-headers==4.3.*
factory-boy==3.3.*
pytest==7.4.*
pytest-django==4.6.*
django-allauth==0.57.*
dj-rest-auth==5.0.*
djangorestframework-simplejwt==5.3.0
